requirements.txt
ccxt==4.4.88
websocket-client==1.7.0
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
click==8.1.7
tabulate==0.9.0
ntplib
flask
flask-socketio
gevent
pandas
numpy
matplotlib
loguru 